backend/src/requirements.txt
blinker==1.6.2
click==8.1.7
ecdsa==0.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyasn1==0.5.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
typed-ast==1.5.5
typing_extensions==4.7.1
Werkzeug==2.3.7
